requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0

# Config & environment
python-dotenv>=1.0.0
pyyaml>=6.0.1

# LangChain / LangGraph
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langgraph>=0.1.0

# Document generation
python-docx>=1.1.0

# Utilities
httpx>=0.27.0
